Interim 9

+++ src/main/java/org/apache/datasketches/kll/KllFloatsSketch.java
@@ -455,8 +455,8 @@ public class KllFloatsSketch extends KllHeapSketch {
     wmem.putByte(PREAMBLE_INTS_BYTE_ADR, (byte) PREAMBLE_INTS_FLOAT);
     wmem.putByte(SER_VER_BYTE_ADR, SERIAL_VERSION_EMPTY_FULL);
     wmem.putByte(FAMILY_BYTE_ADR, (byte) Family.KLL.getID());
-    final byte flags = (byte)
-        ((isLevelZeroSorted() ? LEVEL_ZERO_SORTED_BIT_MASK : 0)
+    final byte flags = (byte) (
+          (isLevelZeroSorted() ? LEVEL_ZERO_SORTED_BIT_MASK : 0)
         | UPDATABLE_BIT_MASK);
     // (leave blank)
     wmem.putByte(FLAGS_BYTE_ADR, flags);
@@ -471,9 +471,9 @@ public class KllFloatsSketch extends KllHeapSketch {
     final int len = getLevelsArray().length;
     wmem.putIntArray(offset, getLevelsArray(), 0, len);
     offset += len * Integer.BYTES;
-    wmem.putDouble(offset, minValue_);
+    wmem.putFloat(offset, minValue_);
     offset += Float.BYTES;
-    wmem.putDouble(offset, maxValue_);
+    wmem.putFloat(offset, maxValue_);
     offset += Float.BYTES;
     wmem.putFloatArray(offset, items_, getLevelsArrayAt(0), getNumRetained());
     return bytes;
